Initial MinGW-w64 support (with CMake) Note: Only the OculusVR make target can be built.  The samples are broken.
|CMake/CMakeLists.txt|
EDID is only built for Linux
Added a file for defining rules for MinGW-w64, this is most useful for cross-compiling on Linux

|Samples/CommonSrc/CMakeLists.txt|
Commented out the DirectX requirement to build for MinGW-w64, this will need to be fixed asap

|LibOVR/Src/Kernel/OVR_ThreadsWinAPI.cpp|
Using the correct bit-size integers when setting the thread's name (dwThreadID)
Thread::SetThreadName is using the method described on MSDN for the RaiseException call
As MinGW-w64 doesn't support Microsoft's structured exception handling, it is not implemented.  A comment has been left on the pertaining lines.

|LibOVR/Src/OVR_Profile.cpp|
GetBaseOVRPath - MinGW-w64 does not provide an implementation of shlobj (that I can find), therefore it is not implemented.  A comment has been left on the pertaining lines.

|LibOVR/Src/OVR_SensorFusion.cpp|
SensorFusion::LoadMagCalibration - MinGW doesn't have sscanf_s, so it is replaced with sscanf

|LibOVR/Src/OVR_Win32_HIDDevice.cpp|
HIDDevice::HIDInitialize - LogText is taking an OVR::String, which needs to expose the internal char *, fixed to use ToCStr( )

|LibOVR/Src/OVR_Win32_HIDDevice.h|
NTSTATUS was not defined, including ntdef.h fixed this issue

|Samples/CommonSrc/Renderer/Render_D3D1X_Device.h|
The header file Windows.h is not playing well on *nix file systems, changed to a lowercase 'w'